+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 #### 1.5.1 (2026-04-22)
 
 - bug fix
+1. 表单中 initialValues 不生效
   修复 ProForm initialValues 默认值配置在表单 prop 存在时不生效的问题
   修复 DialogForm initialValues 默认值配置在表单 prop 存在时不生效的问题
   修复 ProTable initialValues 默认值配置在表单 prop 存在时不生效的问题
+
+2. 表单中 submitter 按钮原生属性不生效
+  修复 ProForm submitter 按钮原生属性不生效的问题
+  修复 DialogForm submitter 按钮原生属性不生效的问题
 
 #### 1.5.0 (2026-04-05)